backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
Pillow==11.0.0
exifread==3.0.0
python-docx==1.1.2
PyPDF2==3.0.1
mutagen==1.47.0
reportlab==4.2.5
numpy==2.1.3
scipy==1.14.1
aiofiles==24.1.0
python-magic==0.4.27
structlog==24.4.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
# Auth
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Payments
stripe==11.3.0
# ML-based detection (CPU-only torch)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
soundfile
